+++ app/Http/Controllers/SettingController.php
@@ -25,28 +25,30 @@ class SettingController extends Controller
      */
     public function update(Request $request, Setting $setting)
     {
-        try {
-            DB::beginTransaction();
 
-            $validate = $request->validate([
-                'app_name' => 'required',
-                'copyright' => 'required',
-                'login_title' => 'required',
-                'keywords' => 'nullable',
-                'description' => 'nullable',
-                'logo' => 'nullable|image|mimes:png,jpg,jpeg,svg|max:512'
-            ], [
-                'app_name.required' => 'Nama aplikasi wajib diisi',
-                'copyright.required' => 'Copyright wajib diisi',
-                'login_title.required' => 'Judul login wajib diisi',
-                'logo.image' => 'File logo harus berupa gambar',
-                'logo.mimes' => 'Format logo harus png, jpg, jpeg, atau svg',
-                'logo.max' => 'Ukuran logo tidak boleh lebih dari 512 KB',
-            ]);
+        $validate = $request->validate([
+            'app_name' => 'required',
+            'copyright' => 'required',
+            'login_title' => 'required',
+            'keywords' => 'nullable',
+            'description' => 'nullable',
+            'logo' => 'nullable|image|mimes:png,jpg,jpeg,svg|max:512'
+        ], [
+            'app_name.required' => 'Nama aplikasi wajib diisi',
+            'copyright.required' => 'Copyright wajib diisi',
+            'login_title.required' => 'Judul login wajib diisi',
+            'logo.image' => 'File logo harus berupa gambar',
+            'logo.mimes' => 'Format logo harus png, jpg, jpeg, atau svg',
+            'logo.max' => 'Ukuran logo tidak boleh lebih dari 512 KB',
+        ]);
+
+        DB::beginTransaction();
+
+        try {
 
             if ($request->file('logo')) {
                 $validate['logo'] = $request->file('logo')->store('img', 'public');
-                if ($setting->logo) {
+                if ($setting->logo && Storage::disk('public')->exists($setting->logo)) {
                     Storage::disk('public')->delete($setting->logo);
                 }
             }
